Revert parallelizing flush
I noticed that flush may fail if the write process to the remote is slow
enough (big file for example or just slow speed in general). The channel
on which the client uses to send/receive request/response somehow got
cancelled on some occasion. There is also some cases where the coroutine
frame of one of the ServerConnection functions (on flush case it's
ServerConnection::write) immediately dropped even when the contained
coroutine has not completed yet which leads to heap-use-after-free. This
is indicated by the socket read operation returns with EFAULT error
code. ASAN can detect this as well. The behavior is unpredictable and
depends on the number of the max buffer size of the channel.